requirements.txt
Flask==3.1.0
google-genai==1.11.0
python-dotenv==1.0.1
markdown==3.7
